+++ src/Adapter/SQL/CollectionTable.php
@@ -143,14 +143,14 @@ final class CollectionTable
             ->unsorted()
             ->map(static fn($entity) => Query\Insert::into(
                 $table,
-                new Row(
-                    new Row\Value(
+                Row::new(
+                    Row\Value::of(
                         Column\Name::of('aggregateId')->in($table),
                         $id->value(),
                     ),
                     ...$entity
                         ->properties()
-                        ->map(static fn($property) => new Row\Value(
+                        ->map(static fn($property) => Row\Value::of(
                             Column\Name::of($property->name())->in($table),
                             $property->value(),
                         ))