+++ src/Adapter/SQL/EntityTable.php
@@ -114,13 +114,13 @@ final class EntityTable
 
         return Query\Insert::into(
             $table,
-            new Row(
-                new Row\Value(
+            Row::new(
+                Row\Value::of(
                     Column\Name::of('aggregateId')->in($table),
                     $id->value(),
                 ),
                 ...$properties
-                    ->map(static fn($property) => new Row\Value(
+                    ->map(static fn($property) => Row\Value::of(
                         Column\Name::of($property->name())->in($table),
                         $property->value(),
                     ))
@@ -143,9 +143,9 @@ final class EntityTable
             ->map(
                 fn($properties) => Update::set(
                     $this->name,
-                    new Row(
+                    Row::new(
                         ...$properties
-                            ->map(fn($property) => new Row\Value(
+                            ->map(fn($property) => Row\Value::of(
                                 Column\Name::of($property->name())->in($this->name),
                                 $property->value(),
                             ))